requirements.txt
autograd==1.5
Bottleneck @ file:///C:/ci/bottleneck_1648010904582/work
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
click==8.1.3
colorama==0.4.6
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
dill==0.3.6
et-xmlfile==1.1.0
fonttools==4.25.0
future==0.18.3
joblib==1.2.0
kiwisolver @ file:///C:/ci/kiwisolver_1653292407425/work
Lifetimes==0.11.3
matplotlib @ file:///C:/ci/matplotlib-suite_1647423638658/work
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
munkres==1.1.4
nltk==3.8.1
numexpr @ file:///C:/ci/numexpr_1640704337920/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1652784039997/work
openpyxl==3.1.2
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas @ file:///C:/ci/pandas_1650373458095/work
Pillow==9.0.1
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///C:/Windows/TEMP/abs_90eacd4e-8eff-491e-b26e-f707eba2cbe1ujvbhqz1/croots/recipe/pytz_1654762631027/work
regex==2022.10.31
scikit-learn==1.2.2
scipy @ file:///C:/ci/scipy_1641555170412/work
seaborn @ file:///tmp/build/80754af9/seaborn_1629307859561/work
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
textblob==0.17.1
threadpoolctl==3.1.0
tornado @ file:///C:/ci/tornado_1606924294691/work
tqdm==4.65.0
wincertstore==0.2
wordcloud==1.8.2.2
